requirements.txt
aiohttp==3.9.3
APScheduler==3.10.4
google-api-python-client==2.126.0
google-auth==2.28.1
google-auth-oauthlib==1.2.0
librosa==0.10.1
moviepy==1.0.3
numpy==1.26.4
openai==0.28.1
pydantic==1.10.14
torch==2.2.1
transformers==4.39.3
yt-dlp==2024.3.10
